refactor: マニフェストモジュールを作成
