+++ build_util/make_docs.py
@@ -6,7 +6,7 @@ from voicevox_engine.core.core_adapter import CoreAdapter
 from voicevox_engine.core.core_initializer import CoreManager
 from voicevox_engine.dev.core.mock import MockCoreWrapper
 from voicevox_engine.dev.tts_engine.mock import MockTTSEngine
-from voicevox_engine.engine_manifest import load_manifest
+from voicevox_engine.engine_manifest.engine_manifest import load_manifest
 from voicevox_engine.preset.Preset import PresetManager
 from voicevox_engine.setting.Setting import USER_SETTING_PATH, SettingHandler
 from voicevox_engine.tts_pipeline.tts_engine import TTSEngineManager
