+++ voicevox_engine/app/application.py
@@ -17,7 +17,7 @@ from voicevox_engine.app.routers.tts_pipeline import generate_tts_pipeline_route
 from voicevox_engine.app.routers.user_dict import generate_user_dict_router
 from voicevox_engine.cancellable_engine import CancellableEngine
 from voicevox_engine.core.core_initializer import CoreManager
-from voicevox_engine.engine_manifest import EngineManifest
+from voicevox_engine.engine_manifest.engine_manifest import EngineManifest
 from voicevox_engine.library_manager import LibraryManager
 from voicevox_engine.metas.MetasStore import MetasStore
 from voicevox_engine.preset.Preset import PresetManager
